Refactored printing of error messages
- Display logic now resides in ProgramErrors. Removing the duplication of print logic in DisplayFormat and TokenSeq.
- Removed module specific error messages from ProgramErrors, leaving only the custom StandardErrors to avoid deeply nested namespaces.
- Removed ErrorPass module and incorporated it into the TokenSeq class because it is not reusable or big enough to be extracted.
